Refactor code into react-devtools-shared
